requirements.txt
PySide6>=6.7
numpy>=1.26
Pillow>=10.0
scipy>=1.11
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
